enhance create-datavolume-task with ability to create datasources
the task will be able to create datasource and wait for it to finish.
The task was not renamed, because this is mostly mean as a fix for
release-v0.3 branch. New version of create-datavolume task will be
rewritten to go language and will have support for both datavolumes
and datasources

Signed-off-by: Karel Šimon <ksimon@redhat.com>