feat: add modify-windows-iso-file task
this tasks modifies windows iso and replaces prompt bootloader with
no-prompt bootloader. This helps during automation of win 11
installation.

We can't do any modifications inside iso, because by default win iso is
in UDF format and read only. To extract iso the guestfish is used. After
extraction, bootloader is replaced with no-prompt one. The extracted files are then
packed back by xorriso. The result iso is then converted to raw format
and stored back to original pvc as a replacement of original iso.

Signed-off-by: Karel Šimon <ksimon@redhat.com>