feat: add delete PVC option for modify-data-object task
This PR adds ability to delete PVC via modify-data-object task.
This will help during efi pipeline, because when the imported PVC is
mounted to modify-windows-iso pod, it can't be deleted, because the PVC
contains finalizer to prevent dataloss when mounted to pod. We have to
include the delete pvc step into efi pipeline, so the pvc can be moved
to terminating state and deleted when modify-windows-iso pod is deleted

Signed-off-by: Karel Šimon <ksimon@redhat.com>