feat: do not exit with error when tekton result folder does not exists in disk-uploader
Disk uploader binary can be triggered via Tekton Tasks or via OCP UI in
disk upload feature. When UI triggers the binary, binary is looking for
Tekton result folder, which does not exist and ends with error.
If the Tekton result folder is not found, the binary will just exit with
0.

Signed-off-by: Karel Simon <ksimon@redhat.com>